docs for tests in java, python, cpp

+++ unittests/cpp/solution_0.1/lib/divide.cpp
@@ -1,3 +1,3 @@
 #include "divide.h"
 
-double divide(double a, double b) { return a / b; }
+int divide(int a, int b) { return a / b; }
